+++ src/dataScipy/DataScience.py
@@ -3,6 +3,7 @@ import matplotlib.pyplot as plt
 from scipy.stats import t
 import scipy.stats as stats
 from pandas import DataFrame
+import pandas as pd
 from sklearn import preprocessing
 import statsmodels.api as sm
 
@@ -15,6 +16,10 @@ def calculate_pvalue(residuals):
     Calculate the p-value for the residuals.
     :param residuals: the residuals (list)
     :return: the p-value (float)
+
+    :Example:
+    >>> residuals = list of residuals
+    >>> p_value = calculate_pvalue(residuals)
     """
     t_dist = abs(np.mean(residuals)) / (np.std(residuals, ddof=1) / np.sqrt(len(residuals)))
     p_value = (1 - t.cdf(t_dist, len(residuals) - 1, loc=0, scale=1)) * 2
@@ -36,6 +41,11 @@ def make_length_equal_to_compare(realVals, forecastVals):
     :param realVals: the real values (list)
     :param forecastVals: the forecast values (list)
     :return: the real and forecast values with equal length (list, list)    
+
+    :Example:
+    >>> data = list of data values
+    >>> forecasts = list of forecasts
+    >>> realVals, forecastVals = make_length_equal_to_compare(data, forecasts)
     """
     no_None = forecastVals.count(None)
     if no_None != 0:
@@ -52,6 +62,11 @@ def MSE(realVals, forecastVals):
     :param realVals: the real values (list)
     :param forecastVals: the forecast values (list)
     :return: the mean squared error (float)
+
+    :Example:
+    >>> data = list of data values
+    >>> forecasts = list of forecasts
+    >>> error = MSE(data, forecasts)
     """
     realVals, forecastVals = make_length_equal_to_compare(realVals, forecastVals)
 
@@ -65,6 +80,11 @@ def RMSE(realVals, forecastVals):
     :param realVals: the real values (list)
     :param forecastVals: the forecast values (list)
     :return: the root mean squared error (float)
+
+    :Example:
+    >>> data = list of data values
+    >>> forecasts = list of forecasts
+    >>> error = RMSE(data, forecasts)
     """ 
     realVals, forecastVals = make_length_equal_to_compare(realVals, forecastVals)
 
@@ -78,6 +98,11 @@ def MAE(realVals, forecastVals):
     :param realVals: the real values (list)
     :param forecastVals: the forecast values (list)
     :return: the mean absolute error (float)
+
+    :Example:
+    >>> data = list of data values
+    >>> forecasts = list of forecasts
+    >>> error = MAE(data, forecasts)
     """
     realVals, forecastVals = make_length_equal_to_compare(realVals, forecastVals)
 
@@ -91,6 +116,11 @@ def MAPE(realVals, forecastVals):
     :param realVals: the real values (list)
     :param forecastVals: the forecast values (list)
     :return: the mean absolute percentage error (float)
+
+    :Example:
+    >>> data = list of data values
+    >>> forecasts = list of forecasts
+    >>> error = MAPE(data, forecasts)
     """
     realVals, forecastVals = make_length_equal_to_compare(realVals, forecastVals)
 
@@ -105,6 +135,11 @@ def calculate_Error(error_method, realVals, forecastVals):
     :param realVals: the real values (list)
     :param forecastVals: the forecast values (list)
     :return: the error (float)
+
+    :Example:
+    >>> data = list of data values
+    >>> forecasts = list of forecasts
+    >>> error = calculate_Error(MSE, data, forecasts)
     """
     realVals, forecastVals = make_length_equal_to_compare(realVals, forecastVals)
     
@@ -118,6 +153,11 @@ def get_all_residuals(realVals, forecastVals):
     :param realVals: the real values (list)
     :param forecastVals: the forecast values (list)
     :return: the errors (list)
+
+    :Example:
+    >>> data = list of data values
+    >>> forecasts = list of forecasts
+    >>> residuals = get_all_residuals(data, forecasts)
     """
     realVals, forecastVals = make_length_equal_to_compare(realVals, forecastVals)
 
@@ -130,7 +170,14 @@ def print_error_summary(data, forecasts, **error_method):
     Print the error summary for the forecasts.
     :param data: the time series
     :param forecasts: the forecasts
+    :param error_method: the error methods to use (function)
+
+    :Example:
+    >>> data = list of data values
+    >>> forecasts = list of forecasts
+    >>> print_error_summary(data, forecasts, MSE=MSE, RMSE=RMSE, MAE=MAE, MAPE=MAPE)
     """
+
     print('Error Summary')
     print('-------------')
     for method, error in error_method.items():
@@ -255,25 +302,11 @@ def plot_normal_of_normalized_resids(residuals):
 
 
 
-def plot_autocorrelation_of_normalized_resids(residuals):
+def show_all_normalized_resids_plots(residuals):
     """
-    Plot the autocorrelation of the residuals.
+    Show all the normalized residuals plots.
     :param residuals: the residuals (list)
     """
-    
-    residual_array = np.array(residuals)
-    normalized_resids = residual_array - np.array(residual_array.mean()) / residual_array.std()
-
-    plt.figure(figsize=(6, 4))
-    sm.graphics.tsa.plot_acf(normalized_resids, color='b', ax=plt.gca(), lags=np.arange(1, 30))
-    plt.title("Autocorrelation", loc='center')
-    plt.ylabel("Correlations")
-    plt.xlabel("Lags")
-    plt.show()
-
-
-
-def show_all_normalize_resids_plots(residuals):
 
     # creating dataframe and then standardizing and normalizing
     residual_array = np.array(residuals)
@@ -315,7 +348,11 @@ def show_all_normalize_resids_plots(residuals):
     
     
     # bottom right subplot
-    sm.graphics.tsa.plot_acf(normalized_resids, color='b', ax=ax4, lags=np.arange(1, 30))
+    max_lag = len(residuals)
+    if max_lag > 30:
+        max_lag = 30
+
+    sm.graphics.tsa.plot_acf(normalized_resids, color='b', ax=ax4, lags=np.arange(1, max_lag + 1))
     ax4.set_title("Autocorrelation", loc='center')
     ax4.set_ylabel("Correlations")
     ax4.set_xlabel("Lags")
@@ -337,6 +374,9 @@ def naive_forecast(data):
     Forecast the next value in the time series using the naive method.
     :param data: time series (list)
     :return: naive forecast (float or integer or None)
+
+    :Example:
+    >>> naive_forecast([1, 2, 3, 4, 5])
     """
     if len(data) == 0:
         return None
@@ -351,6 +391,10 @@ def seasonal_forecast(data, k):
     :param data: time series (list)
     :param k: length of the season (integer)
     :return: single value seasonal forecast (float or integer or None)
+
+    :Example:
+    >>> seasonal_forecast([1, 2, 3, 4, 5], 3)
+    >>> seasonal_forecast([1, 2, 3, 4, 5, 6, 7, 8, 9, 10], 5)
     """
     if len(data) < k:
         return None
@@ -366,6 +410,10 @@ def MA_forecast(data, N, step_ahead=1):
     :param n: the number of previous values to use in the moving average (integer)
     :param step_ahead: the number of steps ahead to forecast (integer)
     :return: the forecast (float or integer or None)
+
+    :Example:
+    >>> MA_forecast([1, 2, 3, 4, 5], 3)
+    >>> MA_forecast([1, 2, 3, 4, 5], 3, 2)
     """
 
     if step_ahead < 1:
@@ -392,6 +440,10 @@ def ES_forecast(data, alpha, step_ahead=1):
     :param alpha: the smoothing parameter (float)
     :param step_ahead: the number of steps ahead to forecast (integer)
     :return: the forecast (float or integer or None)
+
+    :Example:
+    >>> ES_forecast([1, 2, 3, 4, 5], 0.5, 1)
+    >>> ES_forecast([1, 2, 3, 4, 5], 0.5, 2)
     """
 
     if (step_ahead < 1):
@@ -420,7 +472,15 @@ def make_forecast(data, forecast_method, **kwargs):
     :param forecast_method: the forecasting method to use (function)
     :param kwargs: the keyword arguments for the forecasting method (parameters for the forecasting method)
     :return: the forecast (float or integer or None)
+
+    :Example:
+    >>> data = list of data values
+    >>> forecast = make_forecast(data, naive_forecast)
+    >>> forecast = make_forecast(data, seasonal_forecast, k=4)
+    >>> forecast = make_forecast(data, MA_forecast, N=4)
+    >>> forecast = make_forecast(data, ES_forecast, alpha=0.5)
     """
+
     try :
         return forecast_method(data, **kwargs)
     except:
@@ -435,6 +495,12 @@ def make_forecast_for_all_data(data, forecast_method, **kwargs):
     :param forecast_method: the forecasting method to use (function)
     :param kwargs: the keyword arguments for the forecasting method (parameters for the forecasting method)
     :return: the forecast list (list)
+
+    :Example:
+    >>> data = list of data values
+    >>> forecasts = make_forecast_for_all_data(data, naive_forecast)
+    >>> forecasts = make_forecast_for_all_data(data, seasonal_forecast, k=4)
+    >>> forecasts = make_forecast_for_all_data(data, MA_forecast, N=4)
     """
     forecasts = [None]
     
@@ -442,6 +508,50 @@ def make_forecast_for_all_data(data, forecast_method, **kwargs):
         forecasts.append(make_forecast(data[:i+1], forecast_method, **kwargs))
 
     return list(forecasts)
+
+
+def compare_forecasts(data, forecasts, **error_metrics):
+    """
+    This function takes in a list of data, a dictionary of forecasts, and
+    a dictionary of error metrics. It returns a dataframe with the error
+    metrics for each forecast.
+    :param data: list of data values (list)
+    :param forecasts: dictionary of forecasts (dictionary)
+    :param error_metrics: dictionary of error metrics (dictionary)
+    :return: dataframe of error metrics (pandas dataframe)
+
+    :Example:
+    >>> data = list of data values
+    >>> forecasts = {
+    ...     'naive': naive_forecast(data),
+    ...     'seasonal': seasonal_forecast(data, k),
+    ...     'MA': MA_forecast(data, N),
+    ...     'ES': ES_forecast(data, alpha)
+    ... }
+    >>> compare_forecasts(data, forecasts, MAE=mean_absolute_error, RMSE=root_mean_squared_error)
+    """
+
+    if len(forecasts) == 0:
+        raise ValueError("No forecast methods were passed to the function.")
+        
+    if len(error_metrics) == 0:
+        raise ValueError("No error metrics were passed to the function.")
+    
+
+    results = {}
+
+    for error_name, error_metric in error_metrics.items():
+        results[error_name] = []
+        for forecast in forecasts.values():
+            error = error_metric(data, forecast)
+            results[error_name].append(error)
+
+    pd.options.display.float_format = '{:.2f}'.format
+
+    df_results = pd.DataFrame(results, index=forecasts.keys())
+    df_results
+    
+    return df_results
 
 
 
@@ -456,18 +566,26 @@ def make_forecast_for_all_data(data, forecast_method, **kwargs):
 
 
 
-def plot_forecasts(data, horizon, forecasts):
+def plot_forecasts(data, horizon, forecasts, time_step=1):
     """
     Plot the forecasts for the time series.
     :param data: the time series (list)
     :param horizon: the time horizon (list)
     :param forecasts: the forecasts (list)
+    :param time_step: the step size for the time axis (integer)
+
+    :Example:
+    >>> data = list of data values
+    >>> horizon = list of time values
+    >>> forecasts = list of forecast values
+    >>> plot_forecasts(data, horizon, forecasts)
+    >>> plot_forecasts(data, horizon, forecasts, time_step=2)
     """
     
     number_of_dots = len(horizon)
 
     plt.figure(figsize=(12, 3))
-    plt.xticks(np.arange(min(horizon), max(horizon)+1, 2))
+    plt.xticks(np.arange(min(horizon), max(horizon)+1, time_step))
     plt.grid()
     plt.plot(horizon, data[:number_of_dots], label='data')
     plt.plot(horizon, forecasts[:number_of_dots], label='forecast')
@@ -475,6 +593,34 @@ def plot_forecasts(data, horizon, forecasts):
     plt.ylabel('Value')
     plt.title('Forecast and Data')
     plt.legend()
+    plt.show()
+
+
+
+def plot_autocorrelation(values):
+    """
+    Plot the autocorrelation of the values.
+    :param residuals: the residuals (list)
+
+    :Example:
+    >>> data = list of data values
+    >>> plot_autocorrelation(data)
+    >>> residuals = list of residuals
+    >>> plot_autocorrelation(residuals)
+    """
+    
+    residual_array = np.array(values)
+    normalized_resids = residual_array - np.array(residual_array.mean()) / residual_array.std()
+
+    max_lag = len(values)
+    if max_lag > 30:
+        max_lag = 31
+
+    plt.figure(figsize=(6, 4)) 
+    sm.graphics.tsa.plot_acf(normalized_resids, color='b', ax=plt.gca(), lags=np.arange(1, max_lag))
+    plt.title("Autocorrelation", loc='center')
+    plt.ylabel("Correlations")
+    plt.xlabel("Lags")
     plt.show()